refactor(mobile): split sync engine
